Hotfix for Monster sometimes bugging out
Re-Baked the nav mesh wider. Too narrow navmesh was causing monster to bug out if Player hugged too close to the wall when running away. Monster tried to check player's last position, and if that pos was too near the wall it was off the grid. Also found another fix for the camera